convert entities' expiration to integer in tests

+++ src/test/java/org/springframework/data/aerospike/convert/MappingAerospikeConverterTests.java
@@ -376,7 +376,7 @@ public class MappingAerospikeConverterTests extends BaseMappingAerospikeConverte
 
         DocumentWithExpirationAnnotationAndPersistenceConstructor document =
             aerospikeConverter.read(DocumentWithExpirationAnnotationAndPersistenceConstructor.class, forRead);
-        assertThat(document.getExpiration()).isCloseTo(TimeUnit.MINUTES.toSeconds(1), Offset.offset(100L));
+        assertThat(document.getExpiration()).isCloseTo((int) TimeUnit.MINUTES.toSeconds(1), Offset.offset(100));
     }
 
     @ParameterizedTest()
